use failpoint callback instead
